Handle memory types and heaps correctly
The code was accidentally mixing up heap indices and type indices in
a number of places, and the type system wasn't being used adequately
to prevent this. This change introduces stronger heap and type index
types and updates all of the examples and tests.

Fix: https://github.com/io7m/jcoronado/issues/6